+++ host/engine.c
@@ -10,6 +10,7 @@
 
 #include "jerryscript.h"
 
+#include <ctype.h>
 #include <stdio.h>
 #include <string.h>